\name{beta.ci}
\alias{beta.ci}
\title{Compute Confidence Intervals for Beta}
\usage{
  beta.ci(beta, theta, conf = 0.95, x)
}
\arguments{
  \item{beta}{An estimate of beta.}

  \item{theta}{An estimate of theta.}

  \item{conf}{Confidence Level.}

  \item{x}{The data}
}
\description{
  Compute Confidence Intervals for Beta
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

