\name{df2.beta}
\alias{df2.beta}
\title{Compute the Value of the 2nd Derivative of Beta}
\usage{
  df2.beta(beta, x)
}
\arguments{
  \item{beta}{An estimate of beta}

  \item{x}{The data}
}
\description{
  Compute the Value of the 2nd Derivative of Beta
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

