\name{df2.beta.theta}
\alias{df2.beta.theta}
\title{Compute the Cross Partial of Weibull's Theta and Beta.}
\usage{
  df2.beta.theta(beta, x)
}
\arguments{
  \item{beta}{An estimate of beta}

  \item{x}{The data}
}
\description{
  Compute the Cross Partial of Weibull's Theta and Beta.
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

