\name{df2.theta}
\alias{df2.theta}
\title{Compute the Value for the Second Derivative of Theta}
\usage{
  df2.theta(beta, x)
}
\arguments{
  \item{beta}{An estimate of beta.}

  \item{x}{The data}
}
\description{
  Compute the Value for the Second Derivative of Theta
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

