\name{LL}
\alias{LL}
\title{Compute the Log Likelihood of the Weibull Distribution}
\usage{
  LL(beta, x)
}
\arguments{
  \item{beta}{An estimate of beta.}

  \item{x}{The data}
}
\description{
  Compute the Log Likelihood of the Weibull Distribution
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

