\name{lr}
\alias{lr}
\title{Compute the Likelihood-Ratio test for a Weibull Distribution}
\usage{
  lr(null, x, digits = 3)
}
\arguments{
  \item{x}{The data.}

  \item{null}{Value under the null hypothesis.}

  \item{digits}{The number of decimal digits.}
}
\description{
  Compute the Likelihood-Ratio test for a Weibull
  Distribution
}
\details{
  This function is a generalized likelihood ratio test that
  allows for differnt specifications of the null
  hypothesis.
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

