\name{newton.raph}
\alias{newton.raph}
\title{Newton Raphson Alogorithm}
\usage{
  newton.raph(f, df, d2f, start, tol = 1e-06, x)
}
\arguments{
  \item{f}{The function that you want to optimize.}

  \item{df}{the 1st derivative of the function.}

  \item{d2f}{the 2nd derivative of the function.}

  \item{start}{The starting value for optimization.}

  \item{tol}{The level of tolerance desired for
  convergence.}

  \item{x}{The data}
}
\description{
  Newton Raphson Alogorithm
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

